Add mention of text selection prevention in `inert` attribute documentation
This commit adds information to the MDN documentation for the `inert` attribute, specifying that it prevents users from selecting text within the element. This behavior is described as similar to using the CSS property `user-select: none` to disable text selection. The commit includes a link to the documentation for the `user-select` CSS property for further reference. Closes #33611.