Remove allow-downloads-without-user-activation sandbox value (#33409)
This was implemented in Chrome at one time, but has been removed. No browser seems to support this currently. There is an open issue with a request to implement this, but that does not seem very active and that does not seem enough to document it on MDN, even as experimental.

- https://groups.google.com/a/chromium.org/g/blink-dev/c/JdAQ6HNoZvk?pli=1
- https://chromestatus.com/feature/5706745674465280
- https://github.com/whatwg/html/issues/3236
- https://github.com/whatwg/html/issues/5513