Common MIME types: Correct .opus type from audio/opus to audio/ogg
The audio/opus type is only used as an RTP payload format.  Because
it’s for streams, not files, it has no associated filename extension.

https://datatracker.ietf.org/doc/html/rfc7587

When storing Opus audio in a file, a container format is used.  The
canonical choice is the Ogg container format, and it’s this usage that
is denoted by the .opus filename extension, associated with the type
audio/ogg, or more specifically ‘audio/ogg; codecs=opus’.

https://datatracker.ietf.org/doc/html/rfc7845
https://wiki.xiph.org/MIME_Types_and_File_Extensions

Also, Ogg is not an acronym and is not written in all-caps.

https://datatracker.ietf.org/doc/html/rfc3533
https://xiph.org/ogg/

Signed-off-by: Anders Kaseorg <andersk@mit.edu>