MoVe user engagment criteria for install prompt to installation from the web section.
I think it's helpful to mention that promoting an app for installation simply means that there will be an install button somewhere. Otherwise the implication could be that the browser will always make this visible and really obvious to the user, which is unfortunately not always the case.. Safari hides installation behind the share menu+add to homescreen on iOS, and file+add to dock on macOS. Meanwhile other browser typically have it in the overflow menu often with a visual cue, and sometimes the address bar too.