Add Function.length example including destructuring patterns
As with rest and default parameters, destructuring patterns in parameter lists can be a little misleading when it comes to function length. Each pattern counts as a single parameter rather than the count including each identifier within the destructuring pattern. This example shows how length is affected by such patterns.