Add Function.length example including destructuring patterns (#34598)
* Add Function.length example including destructuring patterns

As with rest and default parameters, destructuring patterns in parameter lists can be a little misleading when it comes to function length. Each pattern counts as a single parameter rather than the count including each identifier within the destructuring pattern. This example shows how length is affected by such patterns.

* Update files/en-us/web/javascript/reference/global_objects/function/length/index.md

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>

* Add description in content text

* Update files/en-us/web/javascript/reference/global_objects/function/length/index.md

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Joshua Chen <sidachen2003@gmail.com>