add view-transitions example to MDN
the user had to open an external file, do a view source, then click on the external stylesheet link, to see the example code. this puts the code directly in the page and links to relevant selectors.

I couldn't find the module landing page for this, so did not do a see also.