Update Request:credentials property description
The first paragraph states that the `credentials` property "determines whether or not the browser sends credentials with the request". This fails to address an obscure and major characteristic of the property, which is allowing the browser to store any cookies present in `Set-Cookie` response headers.

See https://fetch.spec.whatwg.org/#cors-protocol-examples last example: "The developer of foo.invalid returns, now fetching some data from bar.invalid while including credentials. [...] This also makes any `Set-Cookie` response headers bar.invalid includes fully functional (they are ignored otherwise)."

I believe this change is fundamentally necessary to help developers like myself, who are facing the issue of the user agent not storing a cookie even though it is present in the response headers list and all CORS rules are being followed. Some http clients do this automatically, but for fine control, manually deciding when to set the `credentials` property to 'include' is necessary.