Update Using the Fetch API's "Including credentials" section
The second paragraph of this section offers instructions on how to "control whether or not the browser sends credentials". This fails to address an obscure and major characteristic of the property, which is allowing the browser to store any cookies present in `Set-Cookie` response headers.

See https://fetch.spec.whatwg.org/#cors-protocol-examples last example: "The developer of foo.invalid returns, now fetching some data from bar.invalid while including credentials. [...] This also makes any `Set-Cookie` response headers bar.invalid includes fully functional (they are ignored otherwise)."

I believe this characteristic may be what the doc's author meant with "If the server does set the correct headers, then the response, including credentials, will be delivered to the caller". This is the only instance I could find where credentials are mentioned in the context of the server's response, and not just what can be sent by the user agent.