Native_manifests: pkcs11 paths may be relative
Undocumented support for relative paths existed for a long while, and
became an explicit first-class supported feature in
https://bugzilla.mozilla.org/show_bug.cgi?id=1759162

This fixes the documentation accordingly, to set the correct
expectations.

And prepend "PKCS 11" to "module" to emphasize that this fragment of the
table is about the manifests for the pkcs11 API, to avoid confusion in
case the reader was looking for the native messaging manifest.