Replace 'elements' with 'entries'
When I was reading this page I did a double take seeing the word 'elements' wondering if that referred to having HTML elements in the history API. But the history API has nothing to do with HTML elements. So seems like maybe a word like 'entries' may be a bit better than 'elements'.