Update code for better understanding (#35313)
* Update code for better understanding.md

I think the for...of example, could be better worded, as we can't use for...of in a javascript object. Compared to the for...in example, which also uses the same example "for (variable of object)". Hence I am proposing a change for such that;

for (variable of iterable objects) 

I wrote iterable objects, as on the paragraph directly above this line in the current documentation, iterable objects were explained and also given a hyperlink.

Many thanks.

* Update files/en-us/web/javascript/guide/loops_and_iteration/index.md

---------

Co-authored-by: Joshua Chen <sidachen2003@gmail.com>