Clarify note on customized built-ins support in Safari
The "Using Custom Elements" page lists customized build-ins first in the discussion of types of custom elements, even though it isn't and won't be supported in Safari. It seems like we should mention the cross-browser supported feature first.

Also, the note on this page doesn't actually say anything about the implementation status, but links to the is= attribute, making it harder to see that Safari will not implement the feature.

I replaced the note here with the note from the is= attribute page, which is a lot more clear and informative.