Client-side form validation: Fix input element styling (#35576)
* Use correct invalid class selector when input is in focus

This example doesn't use the constraint validation API and is meant to use `.invalid` class selector instead of `:invalid` pseudo class selector.

* Increase the specificity of input.invalid selector

The `border-color: #900;` declaration wasn't overriding the border styling added in the `input#mail` selector that precedes it because the `input#mail` selector has a higher specificity. By changing the `input.invalid` selector to `input#mail.invalid`, it now has a higher specificity and the styling successfully applies.