fix a11y issue with keyframes animation
Changing the display property in the keyframes animation example causes focus to be dropped onto the body element when opening or closing the dialog. This is harmful for accessibility, since a keyboard and/or screen reader user may become confused when focus is not managed correctly. The dialog may still be animated using CSS keyframes without changing the value of the display property.