Updated webdriver doc
Given the --remote-debugging-port flag is used, the webdriver property will only be true when port 0 is specified (see https://github.com/GoogleChrome/chrome-launcher/blob/main/docs/chrome-flags-for-tools.md#catch-all-automation). 

Updated language to reflect this. It should also be true when the --remote-debugging-pipe flag is used, but I haven't seen this with testing myself yet.