Fix aspect-ratio example
Since the examples (with EmbedLiveSample) are run on a different domain than
MDN, they don't have access to images hosted on MDN. Using an absolute URL
fixes this issue. Since the same image is already published, the copy from MDN
can be removed, and the already in-use variant can be used.