Reflect that GET, HEAD, and POST are always allowed methods
We recently came across the behaviour that even a non-simple POST request is allowed by the browsers, even if POST is not explicitely stated in the A-C-Allow-Methods header. So I looked it up in the specification and found that in fact _all_ GET, HEAD, and POST requests are CORS-safelisted methods (see also fetch-spec https://fetch.spec.whatwg.org/#ref-for-cors-safelisted-method%E2%91%A2).

A google search revealed that I am not alone in being confused by this. This MDN page added to the confusion, as it shows an example allowing explicitely GET and POST by using this header, which does not make sense, as these HTTP methods are always allowed either way.

So I changed the example to be more meaningful and added a short description to the <method> directive.