Clarity change: replace "okay" with "not too hard"
In the original sentence, "These are usually okay to fix, as long as you are familiar with the right tools and know what the error messages mean!", readers might understand 'okay' to mean 'allowable'. The sentence reads like "you probably won't cause a problem if you fix this issue" -- when I think it's actually meant to say "these issues probably aren't too difficult to fix".