refactor(workflows): extract pr-test-legacy
Variables are not passed to workflows that are triggered
by a pull request from a fork, so mdn/content#37766 did
not work as intended, and rari is always used for tests.

To mitigate this, we extract a new legacy test workflow,
that we can disable, or re-enable in case of rari issues.