Update getPermission() to handle "top-level-storage-access" correctly
Querying for `top-level-storage-access` [requires](https://developer.mozilla.org/en-US/docs/Web/API/Document/requestStorageAccessFor) the caller to pass a `requestedOrigin` value in addition to a `name`. This updates the implementation of the `getPermission()` to pass the `origin` of the window to the `query` call so that the required parameters are passed to the `permissions.query` call.