Fix example in ARIA: grid role (#37903)
Replace the incorrect use of `rowIndex` with `colIndex` in the JavaScript example code for initializing the cell datasets.

This error resulted in incorrect `row` and `col` values on the cells, causing the navigation to not work correctly. Pressing PageDown resulted in an endless `do` ... `while` loop that nearly consumed all of my computer's memory over the next few minutes.