Add link to "day example"
I found it difficult to understand what the code example was showing
because it mentioned a day example but it wasn't clear where this example
was. I think it's clearer with a link. (I split the code block in two and
moved the sentence outside of the code block so that I could put a link in
it.)