Removing misleading assertion
The removed phrase incorrectly implies that once can pass a `false` parameter to `preventSilentAccess(..)` to turn it off (ie, allowing silent access). That is not true -- perhaps used to be? -- so removed the mention entirely, since it's not relevant here. The user will always be prompted regardless of whether that method had been called or not.