Clarify extension ID creation
AFAICT nowhere in the docs did it say that the developer can come up with the extension ID on their own, and in several places it's suggested that they _don't_ come up with it on their own (ex: testing locally, uploading manifest V2 extensions with no ID).

For manifest V3, the developer _must_ come up with an ID on their own to submit the extension. I clarified it (I think in the best location) and provided details on the requirements and how to to choose one.