Fix incorrect composite transformation description
The new description in this change for the `add` value now correctly
describes the transforms that happen during the keyframes animation for
the given examples.

When reading through `animation-composition` I noticed an inaccurate
description of how a composite transformation would be applied during
a keyframes animation.

While the description "matched" what the animation displays, it isn't
what happens under the hood when the transformations are applied, and
could confuse readers unfamiliar with chaining transforms. Specifically
the order described was incorrect as transforms are applied right to
left. Additionally the description said that the element was translated
along the "redirected X axis" which is misleading as rotate
transformations do not change an element's X axis orientation.

It was also noticed that the descriptions of the keyframe animations use
the wrong percentages compared to what is actually used in the keyframes
at-rule, so that was fixed as well.

No QA plan.

Fixes: #38537