Moved explanation regarding pseudo-elements being invalid to the proper paragraph.
The explanation about why pseudo-elements are not valid within the `:has()` function was in the wrong paragraph. It was seemingly explaining why the `:has()` pseudo-class cannot be nested, but more accurately explains why _pseudo-elements_ are not valid with `:has()`.

The CSS Selectors Level 4 spec does not provide an explanation for why the `:has()` pseudo-class can't be nested, but it does explain the pseudo-elements.