Remove erroneous line numbers from link to voice-change-o-matic
Current link to sample source code for voice-change-o-matic includes line numbers that pertain to the visualization function, whereas the distortion-related code relevant to this page is mostly in different sections of the code. Generalizing the link to the entire file rather than specific lines as there are 3 different section of the file that contain relevant code (lines 23-44, 69-84, and 185-216).