Update index.md (#39772)
* Update index.md

The description for "Inclusion criteria" has a singular article with a plural noun: "an in-depth criteria." We could drop the article, but the sentence is still incomplete, as the subject is implied ("Provides…") rather than stated, as in the other examples (e.g., "This section covers," etc.). It seemed best to copy the introduction from "Inclusion criteria, which fixes both issues and keeps it consistent.

* Update index.md

Changed indent and removed clauser per wbamberg's comment. 

Under "Learn web development writing guidelines," I also corrected "this articles" to "this article" and removed the hyperlink on "Learn web development" in the first sentence.

I think it's confusing to include the link here because some folks will click "Learn web development" instead of "Learn web development writing guidelines" and end up in the wrong place.

* Update index.md

Added quotation marks for clarity.