Rebuild of Model element PR from https://github.com/mdn/content/pull/39710
This branch is a clean rebuild of the substantive content about the `HTML Model element` from the earlier PR at
https://github.com/mdn/content/pull/39710
where I had alterations to a .lock file.
Since I'm not familiar enough with Git to revert one single file in the overall process, I thought it would be best to start fresh. Apologies for the clutter!