Transfer-Encoding Header: Remove mention of `trailers` value (#40376)
* remove mention of `trailers` value

`trailers` is only valid for the `TE` header. `Transfer-Encoding: trailers` is not a thing. There already is a (correct) note on https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Headers/TE, so this just needs to be removed.

The PR that introduced this (https://github.com/mdn/content/pull/22032) correctly cites the RFC for `TE`, but then mixes up TE and Transfer-Encoding.

* - Remover trailers see also link

---------

Co-authored-by: Hamish Willee <hamishwillee@gmail.com>