Clarify the meaning of absolute units length
These units are named `pixels` and `centimeters`, but it is important to
call out that those are "CSS" pixels and centimeters,  and  that they
_don't_ match usual definitions of things.

Call out explicitly that `1px` != pixel and that `1cm` != centimeter.

I explicitly decided _not_ to mention angular measures, as my goal here
is to give correct intuition (or rather, correct misleading one), rather
than being fully pedantic (for which we have a spec).