Update HTMLLabelElement 'form' property (#40842)
* Update HTMLLabelElement 'form' property

Clarified the definition of the 'form' property to include all form-associated elements.

the form attribute property page is in the works, so may appear red. see https://github.com/mdn/content/pull/40760/

* Clarify form property description in HTMLLabelElement

* Update index.md

Updated the description of the `form` property in the HTMLLabelElement documentation to clarify its relationship with form-associated elements and removed unnecessary notes.

---------

Co-authored-by: Joshua Chen <sidachen2003@gmail.com>