Update index.md of using textures in webgl
A misleading phrase "Because WebGL now requires textures to be loaded from secure contexts" is removed from the document in this PR. It is true that "Loading of WebGL textures is subject to cross-domain access controls", but to make the context secure the data shall be delivered over an authenticated and encrypted channel, i.e,, requires the use of an HTTPS protocol schema. However, there is no mandatory requirement for webgl applications to load textures from secure context. Cross origin requests are also supported for an HTTP protocol. The documentation should be free from misleading claims.  Besides, it is easier to set up a local http server for webgl-example tutorial experiments. I examined the use of an unencrypted http server (Python http.server) with samples of the webgl-examples tutorial and confirm that the samples1-8 work as expected.