Replace short-note by a detailed section to clarify some vars scope inconsistency (#42164)
* Improve the accuracy of a note about scope in JS

The previous note was stating that loops and conditional block scope behaves differently than the function scope, which I think is only true when declaring variables with `var`.

* Replace  short-note by a detailed section to clarify some vars scope inconsistency

The content of this commit is written by @chrisdavidmills, it is originated from this [comment](https://github.com/mdn/content/pull/42164#pullrequestreview-3524303530).

* Apply suggestions from code review

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>

* Apply a missed suggestion by code-review

Clarified the explanation of ReferenceError and variable scope.

* A few fixes

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Chris Mills <chrisdavidmills@gmail.com>