Fix 'duration' parameter description
Since https://github.com/WebAudio/web-audio-api/pull/1681 (in 2018) the `duration` argument is specced [1] as being the duration of the buffer's content to be played, and not necessarily the one of the AudioContext's clock. So if the source node has a `playbackRate` that is not `1`, using this parameter will not be the same as using `stop(duration)`.  

I based the wordings on the specs text, but I don't mind another formulation altogether.

Might also be noted that there is an interop issue here, where on Firefox does follow the specs. I'll open issues to get this sorted out though.

[1]: https://webaudio.github.io/web-audio-api/#dom-audiobuffersourcenode-start-when-offset-duration-duration