Update `view-timeline-name` property (#42114)
* Update `view-timeline-name` property

Updated the description and syntax for the `view-timeline-name` property to clarify its usage and allowed values. include multiple and globalss

* Fallback for unsupported view-timeline-name

Added a fallback message for unsupported browsers.

* Clean up hidden block

Removed unnecessary code block for hidden CSS.

* typo

* Add padding to unsupported view-timeline-name message

* Apply suggestions from code review

Co-authored-by: Dipika Bhattacharya <dipika@foss-community.org>

* Update description of view-timeline-name property

Clarified the usage and functionality of the view-timeline-name property, including its relationship with the view progress timeline and the shorthand property.

---------

Co-authored-by: Dipika Bhattacharya <dipika@foss-community.org>