Fix display toggle in playground not changing value
The [`display: flow-root` playground](https://developer.mozilla.org/en-US/docs/Web/CSS/Guides/Display/Multi-keyword_syntax#display_block_flow-root_and_display_inline_flow-root) has a `<select>` toggle, which is supposed to change the `display` value of some elements to demonstrate the effects of the `flow-root` value.

Currently, its value is read only once when the playground JS first executed, which results in `display: block` always getting set on the elements. This change makes it so that the value is read after each change of the value, resulting in correct setting of the `display` value.